Handle EAGAIN returned by write()
We use mbuffer for zfs send/recv on both sides. When the receiver is
stalled and does not accept any more data, write() on the sender returns
EAGAIN. It is not clear why, according to man 2 write, this can happen
only on nonblocking sockets, which mbuffer does not seem to use.

It seems to work if this error is handled. write() returns EAGAIN
approximately every 10 seconds for as long as the receiver is stalled
and does not accept data.